ESQL: Move a test util out of production code
This moves a testing utility out of the production code path. It had to
be in the production path a while ago, but that's since been fixed and
now it really belongs under src/test/java.
